+++ tests/__init__.py
@@ -4,7 +4,7 @@ import numpy as np
 import pandas as pd
 import neurokit as nk
 
-#from test_bio import *
+from test_bio import *
 from test_statistics import *
 from test_miscellaneous import *